Жизненно важные изменения
Перенес декоратор check_auth
Переделал тест на авторизованность пользователя
Тесты на проект (create, like)

+++ app/project/routes/__init__.py
@@ -4,9 +4,9 @@ from flask import json, render_template, request, jsonify, redirect, current_app
 from flask_classy import FlaskView, route
 from app.project import bp
 from app.project.responses import ProjectResponses
-from app.decorators import check_auth
+from app.auth.decorators import check_auth
 from app.project.decorators import verify_authorship
-from app.models import Project, User, FavoriteProject
+from app.models import Project as ProjectModel, User, FavoriteProject
 from app.auth.utils import get_auth_instance
 from sqlalchemy.sql.functions import func 
 
@@ -17,23 +17,25 @@ class Project(FlaskView):
         id, claims = get_auth_instance().get_current_user_data_from_token()
         per_page, page = 20, request.args.get('page', 1, type=int)
 
-        projects = Project.query.with_entities(Project).order_by(db.desc(Project.create_at))\
+        projects = ProjectModel.query.with_entities(ProjectModel).order_by(db.desc(ProjectModel.create_at))\
             .paginate(page, per_page)
 
         return render_template('project/index.html', projects=projects, user=claims)
 
     @check_auth
     @route('/create/', methods=['GET'])
-    def create():
+    def create(self):
         id, claims = get_auth_instance().get_current_user_data_from_token()
-        return render_template('project/creator.html', user=claims), 200
+        # return render_template('project/creator.html', user=claims), 200
+        return "шаблон"
 
     @route('/<int:project_id>/')
     def get_project(self, project_id):
         id, claims = get_auth_instance().get_current_user_data_from_token()
 
-        project = Project.query.get(project_id)
-        return render_template('project/index.html', project=project, user=claims)
+        project = ProjectModel.query.get(project_id)
+        # return render_template('project/index.html', project=project, user=claims)
+        return "шаблон"
 
     @check_auth
     @request_is_json(error_message=ProjectResponses.BAD_DATA_TYPE, error_code=400)
@@ -42,17 +44,18 @@ class Project(FlaskView):
         id, claims = get_auth_instance().get_current_user_data_from_token()
         project_id = request.json.get('project_id')
 
-        project = Project.query.get(project_id)
+        project = ProjectModel.query.get(project_id)
 
         if project is None:
             return jsonify(ProjectResponses.BAD_PROJECT_ID_DATA), 400
 
         liked_project = FavoriteProject(user_id=id, project_id=project_id)
+        check_liked_project = FavoriteProject.query.get((id, project_id))
 
-        try:
+        if check_liked_project is None:
             db.session.add(liked_project)
             active = True
-        except:
+        else:
             db.session.delete(
                 FavoriteProject.query.get((liked_project.user_id, liked_project.project_id)))
             active = False
@@ -63,8 +66,8 @@ class Project(FlaskView):
 
     @route('/status_favorites/', methods=['POST'])
     def status_favorites(self):
-        projects = Project.query.with_entities(Project.id, func.count(FavoriteProject.project_id).label('count'))\
-            .filter(Project.id == FavoriteProject.project_id)\
+        projects = ProjectModel.query.with_entities(Project.id, func.count(FavoriteProject.project_id).label('count'))\
+            .filter(ProjectModel.id == FavoriteProject.project_id)\
             .group_by(FavoriteProject.project_id).all()
 
         return jsonify({"status": "success", "projects": projects}), 200
@@ -75,7 +78,7 @@ class Project(FlaskView):
     @route('/remove/', methods=['POST'])
     def remove():
         project_id = request.json.get('project_id')
-        project = Project.query.get(project_id)
+        project = ProjectModel.query.get(project_id)
 
         if project is None:
             return jsonify(ProjectResponses.BAD_PROJECT_ID_DATA), 400
